Fixed UpdatedPlayerMat (see description) and worked on E3LM-RM-BM variant2
Changed UpdatedPlayerMat back to a URP material, set it to opaque and used the albedo map as the emission instead. Note that the player turns invisible when taking damage because it uses alpha clipping (where things are either completely opaque or transparent) instead of translucent. Translucency causes weird rendering issues and should not be used in this case.